+++ wgpu/src/api/render_bundle_encoder.rs
@@ -93,7 +93,7 @@ impl<'a> RenderBundleEncoder<'a> {
             &buffer_slice.buffer.inner,
             index_format,
             buffer_slice.offset,
-            Some(buffer_slice.size),
+            Some(buffer_slice.size_expect_nonzero()),
         );
     }